requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
filelock==3.12.0
Flask==2.3.1
fonttools==4.39.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pydensecrf @ file:///E:/ColoSeg/pydensecrf-1.0rc2-cp39-cp39-win_amd64.whl
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.29.0
six==1.16.0
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
zipp==3.15.0
